Refactor middleware to separate file
