+++ backend/internal/auth/auth.go
@@ -18,6 +18,7 @@ const (
 	IsProd = false
 )
 
+// SOLID: Interface Segregation Principle :)
 type Service interface {
 	GetAuthCallbackHandler() http.HandlerFunc
 	LogoutHandler() http.HandlerFunc
@@ -27,7 +28,7 @@ type Service interface {
 }
 
 type service struct {
-	store *sessions.CookieStore // used this one instead
+	store *sessions.CookieStore
 }
 
 func NewService(store *sessions.CookieStore) Service {
@@ -64,29 +65,4 @@ func NewAuth() Service {
 
 	service := NewService(store)
 	return service
-}
-
-func (s *service) AuthMiddleware(next http.Handler) http.Handler {
-	loggers.Debug.Println("AuthMiddleware called")
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		loggers.Debug.Println("Retrieving session...")
-		session, err := s.store.Get(r, "session-name")
-		if err != nil || session.Values["userID"] == nil {
-			loggers.Debug.Println("User not logged in")
-
-			// Check if the request is an AJAX request
-			if r.Header.Get("X-Requested-With") == "XMLHttpRequest" {
-				loggers.Debug.Println("AJAX request detected, sending unauthorized status")
-				http.Error(w, "Unauthorized", http.StatusUnauthorized)
-				return
-			}
-
-			// For non-AJAX requests, redirect to login
-			loggers.Debug.Println("Redirecting to login page...")
-			http.Redirect(w, r, "/auth/google", http.StatusSeeOther)
-			return
-		}
-		// User is authenticated; proceed with the request
-		next.ServeHTTP(w, r)
-	})
 }
